+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Performance improvements of DSC Pullserver reports from DscWorkshop
   - precalculation of values from JSON based status reports at insert time 
 - Added 2022 and Win11 to Azure image name table
+- Add new columns to TaggingData table and view of DSC Pullserver database.
 
 ### Bugs
 - Fixing issue with data disks on Azure